backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-chess>=1.999
stockfish>=3.28.0
sqlalchemy>=2.0.36
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
firebase-admin>=7.0.0
python-multipart>=0.0.17
aiosqlite>=0.20.0
asyncpg>=0.29.0
httpx>=0.27.0
email-validator>=2.1.0
